it is now possible to build the project using cmake but the old makefiles are still included
git-svn-id: https://potassco.svn.sourceforge.net/svnroot/potassco/branches/gringo2@233 3e02f596-3131-0410-8352-e04eeaec37cc

+++ lib/src/basicprogramevaluator.cpp
@@ -120,12 +120,12 @@ void BasicProgramEvaluator::evaluate()
 	{
 		int i = propQ_.back();
 		propQ_.pop_back();
-		assert(i < watches_.size());
+		assert(i < (int)watches_.size());
 		for(IntVector::iterator it = watches_[i].begin(); it != watches_[i].end(); it++)
 		{
-			assert(*it < rules_.size());
+			assert(*it < (int)rules_.size());
 			IntVector &rule = rules_[*it];
-			assert(rule[0] < facts_.size());
+			assert(rule[0] < (int)facts_.size());
 			if(!facts_[rule[0]])
 			{
 				if(rule.size() == 2)
